Reorganize tests, and clean up some warns in MOI wrapper
